move config and extractor pkgs out of internal
