+++ pkg/extractor/blob.go
