+++ pkg/extractor/extractor.go
