+++ pkg/extractor/image/cache.go
