+++ pkg/extractor/image/source.go
